refactor: view controller 통합
